Script to deploy and update standalone servers
This has been iteratively getting developed and tested on a handful of VPS instances, with and without a mod.  Nebula support for mod management is maybe something to consider for the future, but right now it should work with any svn or git repo for a mod that the user has non-interactive access to.  Also supports running either Release, FastDebug, or even Debug, with or without a debugger wrapper (lldb/clang and gcc/gdb both supported).